refactor: split DietPlan into sub-components
